==> Final_Scrapping/1_scraping_final.ipynb <==
import time
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.common.keys import Keys 

driver = webdriver.Chrome()
# IMDb URL for movies released in 2024 
#url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31"

#to find only action and animation
#url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31&genres=action"#genre = action
#url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31&genres=adventure"#genre = adventure
#url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31&genres=animation"#genre = animation
#url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31&genres=crime"#genre = crime
url="https://www.imdb.com/search/title/?title_type=feature&release_date=2024-01-01,2024-12-31&genres=history"#genre = history

# Open the URL
driver.get(url)

# Allow the page to load (you can adjust the sleep time if necessary)
time.sleep(3)


# Define a function to click the "Load More" button
def click_load_more():
    try: 
        # Locate the "Load More" button
        load_more_button = driver.find_element(By.XPATH, '//*[@id="__next"]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[2]/div[2]/div/span/button/span')

        # Scroll to the button to make sure it's in view
        ActionChains(driver).move_to_element(load_more_button).perform()

        # Click the button
        load_more_button.click()

        # Wait for new content to load
        time.sleep(3)
                                            
        return True
    except:
        duration = "N/A"
    #except Exception as e:
     #   print("Error clicking 'Load More' button:", e)
        return False

# Keep clicking "Load More" until it's no longer available
while click_load_more():
    print("Clicked 'Load More' button")

print("No more content to load.")

movie_items = driver.find_elements(By.XPATH,'//*[@id="__next"]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[2]/ul/li')

print(f"Found {len(movie_items)} movie items.")

# Initialize lists to store the scraped data
titles = []
ratings = []
votings = []
durations = []
genres = []


#movie_items = driver.find_elements(By.XPATH,'/html/body/div[2]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]')


for movie_item in movie_items:
    try: #/html/body/div[2]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[2]/ul/li[1]/div/div/div/div[1]/div[2]/div[1]/a/h3
        title = movie_item.find_element(By.XPATH, './div/div/div/div[1]/div[2]/div[1]/a/h3').text
    except:
        title = "N/A"
        
    try: 
        rating = movie_item.find_element(By.XPATH, './div/div/div/div[1]/div[2]/span/div/span/span[1]').text  
    except:
        rating = "N/A"

    try:
        voting = movie_item.find_element(By.XPATH, './div/div/div/div[1]/div[2]/span/div/span/span[2]').text
    except:
        voting = "N/A"

    try:
        duration = movie_item.find_element(By.XPATH, './div/div/div/div[1]/div[2]/div[2]/span[2]').text
    except:
        duration = "N/A"
         
    try:#/html/body/div[2]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[1]/section/div/div[6]/div[2]/div/section/button[1]/span
        genre = movie_item.find_element(By.CLASS_NAME,'ipc-accordion__item__content')
        #genre = movie_item.find_element(By.XPATH, '/html/body/div[2]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[1]/section/div/div[6]/div[2]/div/section/button[1]/span').text
    except:
        genre = "N/A"        

   # //*[@id="__next"]/main/div[2]/div[3]/section/section/div/section/section/div[2]/div/section/div[2]/div[1]/section

    titles.append(title)
    ratings.append(rating)
    votings.append(voting)
    durations.append(duration)
    genres.append(genre)


 #except Exception as e:
    # If any element is not found, skip this movie and print an error message 
   # print(f"Error extracting data for a movie: {e}")
  #  continue

# Create a DataFrame using the extracted data
df = pd.DataFrame({
    'Title': titles,
    'Rating': ratings,
    'Votes': votings,
    'Duration': durations,
    'Genre': 'History'
})


    # # Save the DataFrame to a CSV file (optional)
df.to_csv('scrapping_history_imdb_movies_2024.csv', index=False)
driver.quit()

df
