==> Final_Scrapping/4_Data Import to MYSQL DB.ipynb <==
!pip install sqlalchemy pymysql pandas



import pandas as pd
from sqlalchemy import create_engine

# Load the cleaned merged CSV
df = pd.read_csv("new_cleaned_merged5_imdb_2024.csv")

# MySQL connection details
username = "root"
password = "root"
host = "localhost"
port = 3306
database = "project1_imdb"
table_name = "imdb_2024_movies"

# Create SQLAlchemy engine
engine = create_engine(f"mysql+pymysql://{username}:{password}@{host}:{port}/{database}")

# Upload the DataFrame to the new table
try:
    df.to_sql(table_name, con=engine, if_exists="replace", index=False)
    print(f"Data uploaded successfully to table '{table_name}' in database '{database}'")
except Exception as e:
    print("Failed to upload to MySQL:", e)


#To view Datas in DB 
#select count(*) as sn from project1_imdb.imdb_2024_movies
#select genre from project1_imdb.imdb_2024_movies group by genre
